requirements.txt
angr
r2pipe
capstone
pexpect
pywebview
pyelftools
python-ptrace
markdown
scapy
